Add support for "opengl:matching" = "generic" to the ARB backend.  This is only applicable to vertex shader inputs (at least on this backend) and is ignored otherwise.
When the variables are mapped to generic attributes, the "opengl:attribindex" metadata is set to the index that the input was mapped to in the original ShVariableNode representing the corresponding shader input.

The latter requires a mapping between potentially transformed inputs (for example, ShTransformer changes "INOUT" variables to an explicit input and output), so I added a utility class that tracks the mapping of new variables back to their originals.

Currently the system works fine on the ARB backend, but as soon as issue225 is resolved (cloning the shader input variables), it will be necessary to have the "clone" function also optionally add mapping information of the cloned variables back to the originals, or else do this manually in ArbCode::generate(). In either case the mapping stored in the new utility class must map variables from the final transformed ones all the way back to the ones in m_originalShader for the metadata to get set correctly (and be visible to the user).

A GLSL implementation of this is still pending, as GLSL treats bindings and attributes completely differently... Francois, you have any time to look into a way of implementing something similar on GLSL? :)

git-svn-id: svn://svn.code.sf.net/p/libsh/code/sh/trunk@2729 afdca40c-03d6-0310-8ede-e9f093b21075