Optimized ArbCode::updateUniform to be more efficient. It is approximately 2.5x faster now, and there's still room for improvement.
I tested the changes in a variety of cases but as always, it's hard to cover every possibility. I'm on Windows right now, but if someone could run the test suite/examples in Linux and verify that I didn't break anything, that would be appreciated.

One other side note: the function is slightly less efficient if it has to do data conversion (i.e. double -> float or something), although it's still faster than it used to be. However when I was testing this case, I discovered that using non-float uniforms (double in this case) was causing Sh to call "evaluate", which was slowing down the program by 30x or more, to a totally unusable state. Perhaps this is the next thing to look at, depending on how important non-float uniform types are :)

git-svn-id: svn://svn.code.sf.net/p/libsh/code/sh/trunk@3225 afdca40c-03d6-0310-8ede-e9f093b21075