Cast GLSL logical operations (<, >, ==, etc) to statement destination types (ex. float). This resolves issue279, where not having this case would cause the shader to fail to compile.
NOTE: Francois if you don't like this fix for some reason, please improve it :) It seemed like a reasonable place to put it, but I don't know the Glsl backend like you do...

NOTE: This change should also be added to the CHANGELOG, but I'm scared of editting it on Windows... last time I did it totally buggered up the line endings, tabs, etc.

Fixed indentation.

git-svn-id: svn://svn.code.sf.net/p/libsh/code/sh/trunk@3442 afdca40c-03d6-0310-8ede-e9f093b21075